analysis_service/requirements.txt
Django==3.2
djangorestframework==3.12.4
psycopg2==2.8.6
djangorestframework-simplejwt==4.6.0
# Pillow==8.4.0
drf-spectacular==0.20.2
matplotlib==3.5.1
